Added tests. Removed extensions project

+++ src/Crafting/ICraftingCalculator.cs
@@ -7,7 +7,7 @@ namespace GilGoblin.Crafting;
 public interface ICraftingCalculator
 {
     Task<int> CalculateCraftingCostForRecipe(int worldID, int recipeID);
-    Task<(int, float)> CalculateCraftingCostForItem(int worldID, int itemID);
+    Task<(int, int)> CalculateCraftingCostForItem(int worldID, int itemID);
     Task<int> CalculateCraftingCostForIngredients(
         int worldID,
         IEnumerable<CraftIngredientPoco> craftIngredients
